Update the beta installation instructions
The instructions in the README assumed that the Flathub base/default
repository is configured. However, when this is not the case, installing
Fractal from Flathub Beta fails because it tries to pull the GNOME
runtime from Flathub base/default.
This adds instructions to install the Flathub base/default repository,
which is harmless if it is already there.