message_row: Improve message layout
Move the ReadReceiptsList to be side-by-side with the rest of the
MessageRow. This is helpful to prevent message rows from jumping around
in the room history due to updating read receipts.

This commit also changes the max receipts shown on the ReadReceiptsList
component to 4, to make more room to its siblings when in the MessageRow
and moves the timestamp label to be closer to the user's name since it
looks odd closer to the read receipts but not touching the outer
container.

Fixes #1313