timeline: Do not use channels to pass data from streams
Even with a buffer of 100 values, sometimes it is not enough during
startup. That makes the task panic, and no more diffs are received in
the timeline.
Instead of growing the buffer to an arbitrary value that might be
way too much or that might break in some cases, let's just use
glib::SendWeakRef and upgrade them when the diff arrives.