Fix media timeline to load in encrypted rooms
When loading the timeline, the messages retrieved were always filtered by the event type `RoomMessage` and by those with URLs. However, when encryption is used, the event type for all messages are 'RoomEncrypted' and the contents cannot be filtered by URL. The fix changed to filter for events of type `RoomEncrypted` and `RoomMessage` when the room has encryption enabled, and leaving the filter unchanged otherwise.

Fixes #1322.