misc: Reduce conversions of string types
Prefer to use `&OwnedType` rather than `&Type` that needs to create
a new `Arc` with `ToOwned`.
To better keep track of those conversions, prefer `.clone` when cloning
from a reference.
Expose `_string` properties rather than convert the owned type to a
string,
then parse it back to the owned type.