user-page: Port ID elements to GtkCenterBox
This ports the user ID label and copy user ID button from GtkBox
to GtkCenterBox. While the GtkBox approach worked, it required us
to use magic numbers (in this case `46`) to center the label.
GtkCenterBox addresses that with a proper mechanism to center
the label, and place the copy user ID button to the end of the
widget, with a margin-start of 6. Visually, both implementations
should look nearly identical.