session: add a keyboard shortcut to go to an unread room
Pressing ctrl+* switches to a room with
- a mention (actually a notification count)
- or any unread message (a room in bold)
- and distinct from the current one

In case rooms with mentions and unread messages are both present,
mentions are preferred.

Favorites sections are also chosen preferentially.

this enables the following workflow:
repeatedly pressing ctrl+* visits all channels with something to be read,
starting with mentions and favorite rooms.

This feature is present in weechat and nheko (bound to alt+A)