room-history: Fix behaviour on scroll
Try to be smarter about what we are doing when we get signals from the
vadjustment.

We can ignore the upper and page-size signals when we are already
auto-scrolling, since trying to scroll more will just result in even
more signals later.

Only focus the last child once, then rely on scrolling to the end to
make sure that we are at the bottom.