room: Construct timeline right away
A recent change made us wait for the category to be loaded before
constructing the Timeline, to know if we should preload it.

However the timeline property of Room assumes that there is always a
timeline available after the construction, which was not true anymore.

So we construct the Timeline during construction again, and use a
property to preload it when both the category and the matrix timeline
are ready.