session: Make sure a single network check per session is running at a time
If there are a lot of network state changes, we can have several
requests at once. The first will succeed but the other ones will timeout
and we will end up offline although there is nothing wrong with the
network.

To prevent that, we use a Mutex so only one requests is made at a time.

We also schedule another check if the network is fine but the server is
unreachable.