sidebar-section: signal aggregate notification count & highlight changes
Add notification-count, highligh and has-notifications properties to the
sidebar section model and signal on corresponding changes.
Aggregation is done as follows:
  is-read := conjunction_over_all_rooms_in_section(is-read)
  highligh := union_over_rooms_in_section(highligh)
  has-notification := any_over_rooms(has-notifications)
  notification-count := sum_over_rooms_in_section(notification-count)

Functionality of ExpressionList is reused on one of the intermediate
models to populate watches (internally on the underlying model) and signal
change of notification-count or highlight with the room indices.
This implementation recalculates the aggregate expressions via
naive iterations (discarding exactly which room(s)' values  were
changed).

UI state bits are reflected accordingly to consume the signalled changes.

Fixes: #1048