account-settings: Reorganize
The main goal here is to move safety-related settings to a new tab so
we can add more in the future.

Since we do not want more tabs, we have to:

- Move the sessions to a subpage accessible with a button in the
"General" tab. While we are here, we reorder most of this tab.
- Move the settings that were in the "Privacy" section of the "Security"
tab to a new "Safety" tab.
- Rename the "Security" tab to "Encryption", to avoid confusion between
"Security" and "Safety".