window: Avoid showing session page every time that selected session changes
We use different objects for the different stages of initialization of
the session, so the selected item will change several times. However, if
the user switches to another page like the greeter, we don't want to
show the selected session page when it is loaded.

So this separates updating the session in the SessionView, and showing
the proper page for the session. We still show the proper page if we
were showing the loading screen or the SessionView.