hooks: Fix pre-commit script
$Could did not exist, and $Failed was not used, so we replace $Failed by
$Could_not and use it instead of $Could

The conditions for calling check_resources for stage files were wrong,
they were only checking the resource if it was NOT staged, because if
was checking for the empty string instead of the non-empty string.
Since we already have the list of staged files from the check before, we
just check if the resource file is present in it or not.

The name of the file for Blueprint resources changed, so we fix it, and
the `ui-resources.gresource.xml` file doesn't exist anymore so we
remove its check.