greeter: Fix and simplify widget hierarchy
The OfflineBanner cannot be a descendant of the GtkScrolledWindow, it
must be a sibling, because we want it to stay in view.

We remove the AdwClamp that was the child of GtkScrolledWindow, because
the content is already wrapped in other AdwClamps.

Finally, we add a little margin around the GtkScrolledWindow child, so
the content has a bit of space around it when the window is small.