utils: Improve waveform generation
Defaulting to 1 sample per second is not enough for short audio streams.
Besides we never actually provide the duration of the stream to the
function, so we always used the default.

This gets the duration from the file first and uses it to compute the
interval between samples, allowing us to consistently have around 110
samples in the waveform.